requirements.txt
streamlit>=1.31.0
anthropic>=0.40.0
pyyaml>=6.0
jsonschema>=4.20.0
